+++ src/components/TodoList/TodoList.tsx
@@ -1,28 +1,16 @@
 import React, { useMemo } from 'react';
-import classNames from 'classnames';
-import { useAppDispatch, useAppSelector } from '../../app/hooks';
-import { Todo } from '../../types/Todo';
-import { actions as currentTodoActions } from '../../features/currentTodo';
-import { getVisibleTodos } from '../helpers/getVisibleTodos';
+import { useAppSelector } from '../../app/hooks';
+import { getVisibleTodos } from '../../helpers/getVisibleTodos';
+import { TodoItem } from '../TodoItem';
 
 export const TodoList: React.FC = () => {
   const todos = useAppSelector(state => state.todos);
   const filter = useAppSelector(state => state.filter.status);
   const query = useAppSelector(state => state.filter.query);
-  const currentTodo = useAppSelector(state => state.currentTodo);
-  const dispatch = useAppDispatch();
-
-  const setCurrentTodo = (todo: Todo) => {
-    dispatch(currentTodoActions.setTodo(todo));
-  };
 
   const visibleTodos = useMemo(() => (
     getVisibleTodos(filter, todos, query)
   ), [todos, filter, query]);
-
-  const handleButtonClick = (todo: Todo) => {
-    setCurrentTodo(todo);
-  };
 
   return (
     <>
@@ -49,49 +37,7 @@ export const TodoList: React.FC = () => {
 
           <tbody>
             {visibleTodos.map(todo => (
-              <tr
-                data-cy="todo"
-                key={todo.id}
-                className={classNames({
-                  'has-background-info-light': currentTodo?.id === todo.id,
-                })}
-              >
-                <td className="is-vcentered">{todo.id}</td>
-                <td className="is-vcentered">
-                  {todo.completed && (
-                    <span className="icon" data-cy="iconCompleted">
-                      <i className="fas fa-check" />
-                    </span>
-                  )}
-                </td>
-
-                <td className="is-vcentered is-expanded">
-                  <p className={classNames({
-                    'has-text-danger': !todo.completed,
-                    'has-text-success': todo.completed,
-                  })}
-                  >
-                    {todo.title}
-                  </p>
-                </td>
-
-                <td className="has-text-right is-vcentered">
-                  <button
-                    data-cy="selectButton"
-                    className="button"
-                    type="button"
-                    onClick={() => handleButtonClick(todo)}
-                  >
-                    <span className="icon">
-                      <i className={classNames('far', {
-                        'fa-eye': currentTodo?.id !== todo.id,
-                        'fa-eye-slash': currentTodo?.id === todo.id,
-                      })}
-                      />
-                    </span>
-                  </button>
-                </td>
-              </tr>
+              <TodoItem todo={todo} key={todo.id} />
             ))}
           </tbody>
         </table>
