+++ src/types/Status.ts
@@ -1,1 +1,5 @@
-export type Status = 'all' | 'active' | 'completed';
+export enum Status {
+  All = 'all',
+  Active = 'active',
+  Completed = 'completed',
+}
